agent_task4/my_ai_agent/requirements.txt
streamlit
pypdf
openagents
google-genai